subcommand olsun ve halihazırda olup olmamasına baksın
