+++ src/commands/index.ts
@@ -1,6 +1,5 @@
 export * from './api/characters';
 export * from './api/check';
-export * from './api/hide';
 export * from './api/player';
 export * from './api/playtime';
 export * from './api/verify';
